refactor(thread): remove unnecessary module thread
PyCharm debugger fails with:

since module thread is internal for Python.
