+++ sdcm/cdclog_reader_thread.py
@@ -7,9 +7,8 @@ from typing import List, Dict
 
 from sdcm.sct_events import CDCReaderStressEvent, Severity
 from sdcm.utils.common import get_docker_stress_image_name
-from sdcm.utils.thread import DockerBasedStressThread
 from sdcm.utils.docker_utils import RemoteDocker
-from sdcm.stress_thread import format_stress_cmd_error
+from sdcm.stress_thread import format_stress_cmd_error, DockerBasedStressThread
 
 LOGGER = logging.getLogger(__name__)
