+++ sdcm/ndbench_thread.py
@@ -9,9 +9,8 @@ from sdcm.prometheus import nemesis_metrics_obj
 from sdcm.sct_events import NdbenchStressEvent, Severity
 from sdcm.utils.common import FileFollowerThread
 from sdcm.remote import FailuresWatcher
-from sdcm.utils.thread import DockerBasedStressThread
 from sdcm.utils.docker_utils import RemoteDocker
-from sdcm.stress_thread import format_stress_cmd_error
+from sdcm.stress_thread import format_stress_cmd_error, DockerBasedStressThread
 
 LOGGER = logging.getLogger(__name__)
